Remove ReviewUx class and related using directives
The `ReviewUx` class has been completely removed from the `ReviewUX.cs` file, along with its properties and the associated using directives for `Dentizone.Domain.Enums` and `Dentizone.Domain.Interfaces`. This change eliminates the class definition and its functionality from the codebase.

+++ Dentizone.Application/DTOs/Review/CreateReviewDTO.cs
@@ -5,7 +5,7 @@ namespace Dentizone.Application.DTOs.Review
     public class CreateReviewDto
     {
         public string OrderId { get; set; }
-        public decimal Stars { get; set; }
+        public int Stars { get; set; }
         public string Comment { get; set; }
     }
